+++ src/renderer/views/projects/details/DependenciesTable.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import Table, { ColumnsType } from 'antd/es/table';
 import LatestVersionCell from './LatestVersionCell';
 import { ParsedDependency } from '@type/ProjectInfo';
@@ -16,48 +16,28 @@ import { useUnit } from 'effector-react';
 import { packageListStore } from '@renderer/stores/PackageListStore';
 import { createPackage, deletePackage } from '@renderer/effects/PackageEffect';
 import { GetPackagesResult } from '@type/PackageListenerArgs';
-import { openAlert } from '@renderer/components/Alert/Alert';
 import { navigateTo } from '@renderer/effects/MenuEffect';
+import {
+  TabPageConfiguration,
+  TabKey,
+  updateTabPageConfig,
+} from '@renderer/stores/DependenciesTabStore';
 
 interface DependenciesTableProps {
+  tabKey: TabKey;
   dependencies: ParsedDependency[];
   registryUrl: string;
+  pageConfig: TabPageConfiguration;
 }
 
 const DependenciesTable: React.FunctionComponent<DependenciesTableProps> = (
   props,
 ) => {
-  const { dependencies, registryUrl } = props;
+  const { tabKey, dependencies, registryUrl, pageConfig: tabConfig } = props;
 
   const { t } = useTranslation();
 
   const { fetchedPackages } = useUnit(packageListStore);
-
-  useEffect(() => {
-    return createPackage.done.watch(({ result }) => {
-      if (!result) {
-        openAlert(
-          'success',
-          t('project.details.alert.title.dependencyFollowed'),
-        );
-      } else {
-        openAlert(
-          'error',
-          t('project.details.alert.title.dependencyFollowError'),
-          t('project.details.alert.description.dependencyFollowError'),
-        );
-      }
-    });
-  });
-
-  useEffect(() => {
-    return deletePackage.done.watch(() => {
-      openAlert(
-        'success',
-        t('project.details.alert.title.dependencyUnfollowed'),
-      );
-    });
-  });
 
   const dependenciesTableColumns: (
     followedPackages: GetPackagesResult,
@@ -157,8 +137,13 @@ const DependenciesTable: React.FunctionComponent<DependenciesTableProps> = (
         key: dependency.name,
       }))}
       pagination={{
+        current: tabConfig.page,
+        defaultPageSize: tabConfig.pageSize,
         position: ['bottomCenter'],
         showSizeChanger: true,
+        onChange(page, pageSize) {
+          updateTabPageConfig({ tabKey, page, pageSize });
+        },
       }}
     />
   );