+++ webpack.renderer.config.ts
@@ -2,11 +2,29 @@ import type { Configuration } from 'webpack';
 import * as path from 'path';
 import { rules } from './webpack.rules';
 import { rendererPlugins } from './webpack.plugins';
+import { loader as miniCssLoader } from 'mini-css-extract-plugin';
+
+const getStyleLoaderPlugin = () => {
+  if (
+    process.env.ENVIRONMENT !== undefined &&
+    process.env.ENVIRONMENT === 'DEV'
+  ) {
+    return {
+      loader: 'style-loader',
+      options: {
+        attributes: {
+          nonce: 'devOnly',
+        },
+      },
+    };
+  }
+  return { loader: miniCssLoader };
+};
 
 rules.push({
   test: /\.s[ac]ss$/i,
   use: [
-    { loader: 'style-loader' },
+    getStyleLoaderPlugin(),
     { loader: 'css-loader' },
     { loader: 'sass-loader' },
   ],
@@ -30,4 +48,5 @@ export const rendererConfig: Configuration = {
       '@renderer/views': path.resolve(__dirname, 'src/renderer/views'),
     },
   },
+  devtool: 'source-map',
 };